created validation tools and implemented firing rate computation
