+++ spiketoolkit/postprocessing/postprocessinglist.py
@@ -1,3 +1,3 @@
-from .postprocessing_tools import get_unit_waveforms, get_unit_templates, get_unit_max_channels, compute_unit_pca_scores, \
-    get_firing_times_ids, get_quality_metric_data, export_to_phy, set_unit_properties_by_max_channel_properties
+from .postprocessing_tools import get_unit_waveforms, get_unit_templates, get_unit_max_channels, compute_unit_pca_scores,\
+     export_to_phy, set_unit_properties_by_max_channel_properties
