+++ spiketoolkit/validation/__init__.py
@@ -1,1 +1,2 @@
 from .qualitymetrics import compute_unit_SNR, compute_firing_rates
+from .validation_tools import get_firing_times_ids, get_quality_metric_data
